Add copywrite header to new files
